decimal percentage

+++ IndieVisible.Application/Services/TranslationAppService.cs
@@ -425,7 +425,7 @@ namespace IndieVisible.Application.Services
         {
             int totalTranslationsTarget = languageCount * totalTerms;
 
-            int percentage = (100 * translatedCount) / (totalTranslationsTarget == 0 ? 1 : totalTranslationsTarget);
+            double percentage = (100 * translatedCount) / (double)(totalTranslationsTarget == 0 ? 1 : totalTranslationsTarget);
 
             return percentage;
         }
